Implement ingest modes
