+++ packages/cli/CHANGELOG.md
@@ -6,16 +6,17 @@ All notable changes to `@relayburn/cli`.
 
 ### Added
 
-- `burn watch --opencode-stream` now writes direct stream-derived OpenCode records for sessions observed from creation, including completed tool-call events and stream cursor high-watermarks.
+- `burn ingest --watch --opencode-stream` now writes direct stream-derived OpenCode records for sessions observed from creation, including completed tool-call events and stream cursor high-watermarks.
 
 ### Changed
 
+- `burn ingest` now owns one-shot, `--watch`, and `--hook claude` modes, replacing the separate `burn watch` command and old `--runtime claude` hook flag.
 - `burn waste --patterns tool-output-bloat` now sizes oversized tool output via cl100k token counts from user-turn enrichment instead of bytes/4 estimates. Findings on highly compressible payloads (repetitive logs, base64 dumps) may shift below the 15k default threshold.
 - `burn summary --subagent-tree` now reads persisted session relationships, including child-session subagents and fork/continuation annotations.
 
 ### Fixed
 
-- `burn watch --opencode-stream` now preserves stream cursor progress when polling fallback saves file-ingest cursors concurrently.
+- `burn ingest --watch --opencode-stream` now preserves stream cursor progress when polling fallback saves file-ingest cursors concurrently.
 
 ## [0.45.0] - 2026-04-29
