+++ packages/cli/src/index.ts
@@ -4,7 +4,8 @@ export { runContext } from './commands/context.js';
 export { runIngest } from './commands/ingest.js';
 export { runLimits } from './commands/limits.js';
 export { runPlans } from './commands/plans.js';
-export { runWatch, runWatchTick, startWatchLoop } from './commands/watch.js';
+export { runIngestTick, startWatchLoop } from './commands/ingest.js';
+export type { StartWatchLoopOptions, WatchController } from './commands/ingest.js';
 export { runWrapper, runWithAdapter } from './commands/run.js';
 export { parseArgs } from './args.js';
 export { ingestClaudeProjects, ingestCodexSessions, ingestOpencodeSessions, ingestAll } from './ingest.js';
